move options to deals package
