+++ deals/module/deals_test.go
@@ -18,7 +18,6 @@ import (
 	logging "github.com/ipfs/go-log/v2"
 	"github.com/stretchr/testify/require"
 	"github.com/textileio/powergate/deals"
-	"github.com/textileio/powergate/ffs"
 	"github.com/textileio/powergate/tests"
 	"github.com/textileio/powergate/util"
 )
@@ -50,7 +49,7 @@ func TestStore(t *testing.T) {
 			checkErr(t, err)
 			_, pcids, err := storeMultiMiner(m, client, nm, randomBytes(600))
 			checkErr(t, err)
-			pending, err := m.ListStorageDealRecords(ffs.WithOnlyPending(true))
+			pending, err := m.ListStorageDealRecords(deals.WithOnlyPending(true))
 			require.Nil(t, err)
 			require.Len(t, pending, nm)
 			final, err := m.ListStorageDealRecords()
@@ -59,7 +58,7 @@ func TestStore(t *testing.T) {
 			err = waitForDealComplete(client, pcids)
 			checkErr(t, err)
 			time.Sleep(util.AvgBlockTime)
-			pending, err = m.ListStorageDealRecords(ffs.WithOnlyPending(true))
+			pending, err = m.ListStorageDealRecords(deals.WithOnlyPending(true))
 			require.Nil(t, err)
 			require.Empty(t, pending)
 			final, err = m.ListStorageDealRecords()
