+++ deals/types.go
@@ -1,8 +1,6 @@
 package deals
 
 import (
-	"os"
-
 	"github.com/ipfs/go-cid"
 )
 
@@ -63,24 +61,4 @@ type RetrievalDealRecord struct {
 	Addr     string
 	DealInfo RetrievalDealInfo
 	Time     int64
-}
-
-// Config contains configuration for storing deals.
-type Config struct {
-	ImportPath string
-}
-
-// Option sets values on a Config.
-type Option func(*Config) error
-
-// WithImportPath indicates the import path that will be used
-// to store data to later be imported to Lotus.
-func WithImportPath(path string) Option {
-	return func(c *Config) error {
-		if err := os.MkdirAll(path, 0700); err != nil {
-			return err
-		}
-		c.ImportPath = path
-		return nil
-	}
 }
