+++ ffs/interfaces.go
@@ -6,7 +6,6 @@ import (
 	"math/big"
 
 	"github.com/ipfs/go-cid"
-	"github.com/textileio/powergate/deals"
 )
 
 // WalletManager provides access to a Lotus wallet for a Lotus node.
@@ -17,64 +16,6 @@ type WalletManager interface {
 	Balance(context.Context, string) (uint64, error)
 	// SendFil sends fil from one address to another.
 	SendFil(context.Context, string, string, *big.Int) error
-}
-
-// ListDealRecordsConfig specifies the options for DealsManager.List.
-type ListDealRecordsConfig struct {
-	FromAddrs      []string
-	DataCids       []string
-	IncludePending bool
-	OnlyPending    bool
-	Ascending      bool
-}
-
-// ListDealRecordsOption updates a ListDealRecordsConfig.
-type ListDealRecordsOption func(*ListDealRecordsConfig)
-
-// WithFromAddrs limits the results deals initated from the provided wallet addresses.
-func WithFromAddrs(addrs ...string) ListDealRecordsOption {
-	return func(c *ListDealRecordsConfig) {
-		c.FromAddrs = addrs
-	}
-}
-
-// WithDataCids limits the results to deals for the provided data cids.
-func WithDataCids(cids ...string) ListDealRecordsOption {
-	return func(c *ListDealRecordsConfig) {
-		c.DataCids = cids
-	}
-}
-
-// WithIncludePending specifies whether or not to include pending deals in the results.
-// Ignored for ListRetrievalDealRecords.
-func WithIncludePending(includePending bool) ListDealRecordsOption {
-	return func(c *ListDealRecordsConfig) {
-		c.IncludePending = includePending
-	}
-}
-
-// WithOnlyPending specifies whether or not to only include pending deals in the results.
-// Ignored for ListRetrievalDealRecords.
-func WithOnlyPending(onlyPending bool) ListDealRecordsOption {
-	return func(c *ListDealRecordsConfig) {
-		c.OnlyPending = onlyPending
-	}
-}
-
-// WithAscending specifies to sort the results in ascending order.
-// Default is descending order.
-// If pending, records are sorted by timestamp, otherwise records
-// are sorted by activation epoch then timestamp.
-func WithAscending(ascending bool) ListDealRecordsOption {
-	return func(c *ListDealRecordsConfig) {
-		c.Ascending = ascending
-	}
-}
-
-// DealsManager specifies the required deals api.
-type DealsManager interface {
-	ListStorageDealRecords(...ListDealRecordsOption) ([]deals.StorageDealRecord, error)
-	ListRetrievalDealRecords(...ListDealRecordsOption) ([]deals.RetrievalDealRecord, error)
 }
 
 // PaychManager provides access to payment channels.
